Fix chat output

+++ src/my_discord_bot/cogs/_cog_base.py
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 
 
 def check_cooldown_factory(
-    seconds: int = 10,
+    seconds: float = 10,
 ) -> ty.Callable[[discord.Interaction], discord.app_commands.Cooldown | None]:
     """Global cooldown for commands."""
